requirements/base.txt
python-docx
langchain_community
langchain
tqdm
chardet
langchain-text-splitters
langchain_huggingface
chromadb
langchain_ollama
fastapi
uvicorn
streamlit